backend/requirements.txt
Flask==1.1.2
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2-binary>=2.7
Flask-Cors>=3.0
boto3==1.28.84